Consolidate Manifest Handling by Removing KF-Specific Implementation (#470)
* Consolidate Manifest Handling by Removing KF-Specific Implementation

* Removing unused classes and dependecies from the code
