+++ score-client/pom.xml
@@ -147,13 +147,6 @@ ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF S
       <version>${guava.version}</version>
     </dependency>
 
-    <!-- jackson-dataformat-csv -->
-    <dependency>
-      <groupId>com.fasterxml.jackson.dataformat</groupId>
-      <artifactId>jackson-dataformat-csv</artifactId>
-      <version>${fasterxml-jackson.version}</version>
-    </dependency>
-
     <!-- SSL -->
     <dependency>
       <groupId>org.bouncycastle</groupId>
