+++ score-client/src/main/java/bio/overture/score/client/manifest/UploadManifestReader.java
@@ -15,12 +15,11 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package bio.overture.score.client.manifest.icgc;
+package bio.overture.score.client.manifest;
 
 import static com.google.common.base.Preconditions.checkState;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
-import bio.overture.score.client.manifest.UploadManifest;
 import bio.overture.score.client.manifest.UploadManifest.ManifestEntry;
 import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
