--- score-client/src/main/java/bio/overture/score/client/util/CsvParser.java
@@ -1,24 +1,0 @@
-package bio.overture.score.client.util;
-
-import com.fasterxml.jackson.dataformat.csv.CsvMapper;
-import java.io.File;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import lombok.SneakyThrows;
-import lombok.val;
-
-@RequiredArgsConstructor
-public class CsvParser<T> {
-
-  private final Class<T> type;
-  private final Character columnSep;
-
-  @SneakyThrows
-  public List<T> parseFile(File file) {
-    val mapper = new CsvMapper();
-    mapper.addMixIn(type, type);
-    val schema = mapper.schemaFor(type).withHeader().withColumnSeparator(columnSep);
-
-    return mapper.readerFor(type).with(schema).<T>readValues(file).readAll();
-  }
-}
